dynamically load freeze features
